+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Removes metric related configuration options from `ConsoleExporterOptions`.
+  `MetricReaderType`, `PeriodicExporterMetricReaderOptions`, and `Temporality`
+  are now configurable via the `MetricReaderOptions`.
+  ([#2929](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2929))
+
 ## 1.2.0-rc2
 
 Released 2022-Feb-02